@@ -2,14 +2,10 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// @dart = 2.8
-
 import 'dart:math' as math;
 import 'dart:ui' as ui;
 import 'package:flutter/gestures.dart'
     show kMinFlingVelocity, kLongPressTimeout;
-import 'package:flutter/foundation.dart' show kIsWeb;
-import 'package:flutter/rendering.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
@@ -44,7 +40,7 @@ typedef _ContextMenuPreviewBuilderChildless = Widget Function(
 Rect _getRect(GlobalKey globalKey) {
   assert(globalKey.currentContext != null);
   final RenderBox renderBoxContainer =
-      globalKey.currentContext.findRenderObject() as RenderBox;
+      globalKey.currentContext!.findRenderObject()! as RenderBox;
   final Offset containerOffset = renderBoxContainer.localToGlobal(
     renderBoxContainer.paintBounds.topLeft,
   );
@@ -92,7 +88,7 @@ enum _ContextMenuLocation {
 /// Widget build(BuildContext context) {
 ///   return Scaffold(
 ///     body: Center(
-///       child: Container(
+///       child: SizedBox(
 ///         width: 100,
 ///         height: 100,
 ///         child: CupertinoContextMenu(
@@ -131,15 +127,14 @@ class CustomCupertinoContextMenu extends StatefulWidget {
   ///
   /// [child] is required and cannot be null.
   CustomCupertinoContextMenu({
-    Key key,
-    @required this.actions,
-    @required this.child,
+    Key? key,
+    required this.actions,
+    required this.child,
     this.previewBuilder,
     this.enableLongPress = true,
     this.decoyChildAbsorbing = true,
     this.contextMenuChildAbsorbing = true,
-  })  : assert(actions != null && actions.isNotEmpty),
-        assert(child != null),
+  })  : assert(actions.isNotEmpty),
         super(key: key);
 
   /// The widget that can be "opened" with the [CustomCupertinoContextMenu].
@@ -234,7 +229,7 @@ class CustomCupertinoContextMenu extends StatefulWidget {
   /// ```
   ///
   /// {@end-tool}
-  final ContextMenuPreviewBuilder previewBuilder;
+  final ContextMenuPreviewBuilder? previewBuilder;
 
   @override
   _CustomCupertinoContextMenuState createState() =>
@@ -246,10 +241,10 @@ class _CustomCupertinoContextMenuState extends State<CustomCupertinoContextMenu>
   final GlobalKey _childGlobalKey = GlobalKey();
   bool _childHidden = false;
   // Animates the child while it's opening.
-  AnimationController _openController;
-  Rect _decoyChildEndRect;
-  OverlayEntry _lastOverlayEntry;
-  _ContextMenuRoute<void> _route;
+  late AnimationController _openController;
+  Rect? _decoyChildEndRect;
+  OverlayEntry? _lastOverlayEntry;
+  _ContextMenuRoute<void>? _route;
 
   @override
   void initState() {
@@ -302,17 +297,17 @@ class _CustomCupertinoContextMenuState extends State<CustomCupertinoContextMenu>
         sigmaY: 5.0,
       ),
       contextMenuLocation: _contextMenuLocation,
-      previousChildRect: _decoyChildEndRect,
+      previousChildRect: _decoyChildEndRect!,
       builder: (BuildContext context, Animation<double> animation) {
         if (widget.previewBuilder == null) {
           return widget.child;
         }
-        return widget.previewBuilder(context, animation, widget.child);
+        return widget.previewBuilder!(context, animation, widget.child);
       },
       absorbing: widget.contextMenuChildAbsorbing,
     );
-    Navigator.of(context, rootNavigator: true).push<void>(_route);
-    _route.animation.addStatusListener(_routeAnimationStatusListener);
+    Navigator.of(context, rootNavigator: true).push<void>(_route!);
+    _route!.animation!.addStatusListener(_routeAnimationStatusListener);
   }
 
   void _onDecoyAnimationStatusChange(AnimationStatus animationStatus) {
@@ -336,7 +331,7 @@ class _CustomCupertinoContextMenuState extends State<CustomCupertinoContextMenu>
         // because _ContextMenuRoute renders its first frame offscreen.
         // Otherwise there would be a visible flash when nothing is rendered for
         // one frame.
-        SchedulerBinding.instance.addPostFrameCallback((Duration _) {
+        SchedulerBinding.instance!.addPostFrameCallback((Duration _) {
           _lastOverlayEntry?.remove();
           _lastOverlayEntry = null;
           _openController.reset();
@@ -354,13 +349,10 @@ class _CustomCupertinoContextMenuState extends State<CustomCupertinoContextMenu>
     if (status != AnimationStatus.dismissed) {
       return;
     }
-    if (mounted) {
-      setState(() {
-        _childHidden = false;
-      });
-    }
-
-    _route.animation.removeStatusListener(_routeAnimationStatusListener);
+    setState(() {
+      _childHidden = false;
+    });
+    _route!.animation!.removeStatusListener(_routeAnimationStatusListener);
     _route = null;
   }
 
@@ -382,7 +374,7 @@ class _CustomCupertinoContextMenuState extends State<CustomCupertinoContextMenu>
     }
   }
 
-  void _onTapDown(TapDownDetails details) {
+  void _onTapDown(TapDownDetails? details) {
     setState(() {
       _childHidden = true;
     });
@@ -414,7 +406,7 @@ class _CustomCupertinoContextMenuState extends State<CustomCupertinoContextMenu>
         );
       },
     );
-    Overlay.of(context).insert(_lastOverlayEntry);
+    Overlay.of(context)!.insert(_lastOverlayEntry!);
     _openController.forward();
   }
 
@@ -453,18 +445,18 @@ class _CustomCupertinoContextMenuState extends State<CustomCupertinoContextMenu>
 // siblings of the original child.
 class _DecoyChild extends StatefulWidget {
   const _DecoyChild({
-    Key key,
+    Key? key,
     this.beginRect,
-    this.controller,
+    required this.controller,
     this.endRect,
     this.child,
-    this.absorbing,
+    this.absorbing = true,
   }) : super(key: key);
 
-  final Rect beginRect;
+  final Rect? beginRect;
   final AnimationController controller;
-  final Rect endRect;
-  final Widget child;
+  final Rect? endRect;
+  final Widget? child;
   final bool absorbing;
 
   @override
@@ -479,8 +471,8 @@ class _DecoyChildState extends State<_DecoyChild>
   static const Color _masklessColor = Color(0xFFFFFFFF);
 
   final GlobalKey _childGlobalKey = GlobalKey();
-  Animation<Color> _mask;
-  Animation<Rect> _rect;
+  late Animation<Color> _mask;
+  late Animation<Rect?> _rect;
 
   @override
   void initState() {
@@ -495,18 +487,18 @@ class _DecoyChildState extends State<_DecoyChild>
       intervalOff: 0.5,
     );
 
-    final Rect midRect = widget.beginRect.deflate(
-      widget.beginRect.width * (_kOpenScale - 1.0) / 2,
+    final Rect midRect = widget.beginRect!.deflate(
+      widget.beginRect!.width * (_kOpenScale - 1.0) / 2,
     );
-    _rect = TweenSequence<Rect>(<TweenSequenceItem<Rect>>[
-      TweenSequenceItem<Rect>(
+    _rect = TweenSequence<Rect?>(<TweenSequenceItem<Rect?>>[
+      TweenSequenceItem<Rect?>(
         tween: RectTween(
           begin: widget.beginRect,
           end: midRect,
         ).chain(CurveTween(curve: Curves.easeInOutCubic)),
         weight: 1.0,
       ),
-      TweenSequenceItem<Rect>(
+      TweenSequenceItem<Rect?>(
         tween: RectTween(
           begin: midRect,
           end: widget.endRect,
@@ -533,31 +525,23 @@ class _DecoyChildState extends State<_DecoyChild>
     super.dispose();
   }
 
-  Widget _buildAnimation(BuildContext context, Widget child) {
+  Widget _buildAnimation(BuildContext context, Widget? child) {
     final Color color = widget.controller.status == AnimationStatus.reverse
         ? _masklessColor
         : _mask.value;
     return Positioned.fromRect(
-      rect: _rect.value,
-      // TODO(justinmc): When ShaderMask is supported on web, remove this
-      // conditional and use ShaderMask everywhere.
-      // https://github.com/flutter/flutter/issues/52967.
-      child: kIsWeb
-          ? Container(key: _childGlobalKey, child: widget.child)
-          : ShaderMask(
-              key: _childGlobalKey,
-              shaderCallback: (Rect bounds) {
-                return LinearGradient(
-                  begin: Alignment.topLeft,
-                  end: Alignment.bottomRight,
-                  colors: <Color>[color, color],
-                ).createShader(bounds);
-              },
-              child: AbsorbPointer(
-                absorbing: widget.absorbing,
-                child: widget.child,
-              ),
-            ),
+      rect: _rect.value!,
+      child: ShaderMask(
+        key: _childGlobalKey,
+        shaderCallback: (Rect bounds) {
+          return LinearGradient(
+            begin: Alignment.topLeft,
+            end: Alignment.bottomRight,
+            colors: <Color>[color, color],
+          ).createShader(bounds);
+        },
+        child: widget.child,
+      ),
     );
   }
 
@@ -578,14 +562,14 @@ class _DecoyChildState extends State<_DecoyChild>
 class _ContextMenuRoute<T> extends PopupRoute<T> {
   // Build a _ContextMenuRoute.
   _ContextMenuRoute({
-    @required List<Widget> actions,
-    @required _ContextMenuLocation contextMenuLocation,
+    required List<Widget> actions,
+    required _ContextMenuLocation contextMenuLocation,
     this.barrierLabel,
-    _ContextMenuPreviewBuilderChildless builder,
-    ui.ImageFilter filter,
-    Rect previousChildRect,
-    RouteSettings settings,
-    this.absorbing,
+    _ContextMenuPreviewBuilderChildless? builder,
+    ui.ImageFilter? filter,
+    required Rect previousChildRect,
+    RouteSettings? settings,
+    this.absorbing = false,
   })  : assert(actions != null && actions.isNotEmpty),
         assert(contextMenuLocation != null),
         _actions = actions,
@@ -607,15 +591,15 @@ class _ContextMenuRoute<T> extends PopupRoute<T> {
       Duration(milliseconds: 335);
 
   final List<Widget> _actions;
-  final _ContextMenuPreviewBuilderChildless _builder;
+  final _ContextMenuPreviewBuilderChildless? _builder;
   final GlobalKey _childGlobalKey = GlobalKey();
   final _ContextMenuLocation _contextMenuLocation;
   bool _externalOffstage = false;
   bool _internalOffstage = false;
-  Orientation _lastOrientation;
+  Orientation? _lastOrientation;
   // The Rect of the child at the moment that the CupertinoContextMenu opens.
   final Rect _previousChildRect;
-  double _scale = 1.0;
+  double? _scale = 1.0;
   final GlobalKey _sheetGlobalKey = GlobalKey();
 
   static final CurveTween _curve = CurveTween(
@@ -625,17 +609,17 @@ class _ContextMenuRoute<T> extends PopupRoute<T> {
     curve: Curves.easeInBack,
   );
   static final RectTween _rectTween = RectTween();
-  static final Animatable<Rect> _rectAnimatable = _rectTween.chain(_curve);
+  static final Animatable<Rect?> _rectAnimatable = _rectTween.chain(_curve);
   static final RectTween _rectTweenReverse = RectTween();
-  static final Animatable<Rect> _rectAnimatableReverse =
+  static final Animatable<Rect?> _rectAnimatableReverse =
       _rectTweenReverse.chain(
     _curveReverse,
   );
   static final RectTween _sheetRectTween = RectTween();
-  final Animatable<Rect> _sheetRectAnimatable = _sheetRectTween.chain(
+  final Animatable<Rect?> _sheetRectAnimatable = _sheetRectTween.chain(
     _curve,
   );
-  final Animatable<Rect> _sheetRectAnimatableReverse = _sheetRectTween.chain(
+  final Animatable<Rect?> _sheetRectAnimatableReverse = _sheetRectTween.chain(
     _curveReverse,
   );
   static final Tween<double> _sheetScaleTween = Tween<double>();
@@ -648,10 +632,10 @@ class _ContextMenuRoute<T> extends PopupRoute<T> {
     _curveReverse,
   );
   final Tween<double> _opacityTween = Tween<double>(begin: 0.0, end: 1.0);
-  Animation<double> _sheetOpacity;
+  late Animation<double> _sheetOpacity;
 
   @override
-  final String barrierLabel;
+  final String? barrierLabel;
 
   @override
   Color get barrierColor => _kModalBarrierColor;
@@ -686,14 +670,14 @@ class _ContextMenuRoute<T> extends PopupRoute<T> {
         return AlignmentDirectional.topCenter;
       case _ContextMenuLocation.right:
         return AlignmentDirectional.topEnd;
-      default:
+      case _ContextMenuLocation.left:
         return AlignmentDirectional.topStart;
     }
   }
 
   // The place to start the sheetRect animation from.
   static Rect _getSheetRectBegin(
-      Orientation orientation,
+      Orientation? orientation,
       _ContextMenuLocation contextMenuLocation,
       Rect childRect,
       Rect sheetRect) {
@@ -709,7 +693,7 @@ class _ContextMenuRoute<T> extends PopupRoute<T> {
             ? childRect.bottomRight
             : childRect.topRight;
         return (target - Offset(sheetRect.width, 0.0)) & sheetRect.size;
-      default:
+      case _ContextMenuLocation.left:
         final Offset target = orientation == Orientation.portrait
             ? childRect.bottomLeft
             : childRect.topLeft;
@@ -721,7 +705,7 @@ class _ContextMenuRoute<T> extends PopupRoute<T> {
     _scale = scale;
     _opacityTween.end = opacity;
     _sheetOpacity = _opacityTween.animate(CurvedAnimation(
-      parent: animation,
+      parent: animation!,
       curve: const Interval(0.9, 1.0),
     ));
     Navigator.of(context).pop();
@@ -732,7 +716,7 @@ class _ContextMenuRoute<T> extends PopupRoute<T> {
   void _updateTweenRects() {
     final Rect childRect = _scale == null
         ? _getRect(_childGlobalKey)
-        : _getScaledRect(_childGlobalKey, _scale);
+        : _getScaledRect(_childGlobalKey, _scale!);
     _rectTween.begin = _previousChildRect;
     _rectTween.end = childRect;
 
@@ -769,7 +753,7 @@ class _ContextMenuRoute<T> extends PopupRoute<T> {
   }
 
   @override
-  bool didPop(T result) {
+  bool didPop(T? result) {
     _updateTweenRects();
     return super.didPop(result);
   }
@@ -787,7 +771,7 @@ class _ContextMenuRoute<T> extends PopupRoute<T> {
 
     // Render one frame offstage in the final position so that we can take
     // measurements of its layout and then animate to them.
-    SchedulerBinding.instance.addPostFrameCallback((Duration _) {
+    SchedulerBinding.instance!.addPostFrameCallback((Duration _) {
       _updateTweenRects();
       _internalOffstage = false;
       _setOffstageInternally();
@@ -812,8 +796,8 @@ class _ContextMenuRoute<T> extends PopupRoute<T> {
     // buildTransitions as child, the idea being that buildTransitions will
     // animate the entire page into the scene. In the case of _ContextMenuRoute,
     // two individual pieces of the page are animated into the scene in
-    // buildTransitions, and null is returned here.
-    return null;
+    // buildTransitions, and a Container is returned here.
+    return Container();
   }
 
   @override
@@ -828,11 +812,11 @@ class _ContextMenuRoute<T> extends PopupRoute<T> {
         if (!animation.isCompleted) {
           final bool reverse = animation.status == AnimationStatus.reverse;
           final Rect rect = reverse
-              ? _rectAnimatableReverse.evaluate(animation)
-              : _rectAnimatable.evaluate(animation);
+              ? _rectAnimatableReverse.evaluate(animation)!
+              : _rectAnimatable.evaluate(animation)!;
           final Rect sheetRect = reverse
-              ? _sheetRectAnimatableReverse.evaluate(animation)
-              : _sheetRectAnimatable.evaluate(animation);
+              ? _sheetRectAnimatableReverse.evaluate(animation)!
+              : _sheetRectAnimatable.evaluate(animation)!;
           final double sheetScale = reverse
               ? _sheetScaleAnimatableReverse.evaluate(animation)
               : _sheetScaleAnimatable.evaluate(animation);
@@ -857,7 +841,7 @@ class _ContextMenuRoute<T> extends PopupRoute<T> {
               Positioned.fromRect(
                 key: _childGlobalKey,
                 rect: rect,
-                child: _builder(context, animation),
+                child: _builder!(context, animation),
               ),
             ],
           );
@@ -867,13 +851,12 @@ class _ContextMenuRoute<T> extends PopupRoute<T> {
         // in the final position.
         return _ContextMenuRouteStatic(
           actions: _actions,
-          child: _builder(context, animation),
+          child: _builder!(context, animation),
           childGlobalKey: _childGlobalKey,
           contextMenuLocation: _contextMenuLocation,
           onDismiss: _onDismiss,
           orientation: orientation,
           sheetGlobalKey: _sheetGlobalKey,
-          absorbing: absorbing,
         );
       },
     );
@@ -884,27 +867,25 @@ class _ContextMenuRoute<T> extends PopupRoute<T> {
 // animating out.
 class _ContextMenuRouteStatic extends StatefulWidget {
   const _ContextMenuRouteStatic({
-    Key key,
+    Key? key,
     this.actions,
-    @required this.child,
+    required this.child,
     this.childGlobalKey,
-    @required this.contextMenuLocation,
+    required this.contextMenuLocation,
     this.onDismiss,
-    @required this.orientation,
+    required this.orientation,
     this.sheetGlobalKey,
-    this.absorbing,
   })  : assert(contextMenuLocation != null),
         assert(orientation != null),
         super(key: key);
 
-  final List<Widget> actions;
+  final List<Widget>? actions;
   final Widget child;
-  final GlobalKey childGlobalKey;
+  final GlobalKey? childGlobalKey;
   final _ContextMenuLocation contextMenuLocation;
-  final _DismissCallback onDismiss;
+  final _DismissCallback? onDismiss;
   final Orientation orientation;
-  final GlobalKey sheetGlobalKey;
-  final bool absorbing;
+  final GlobalKey? sheetGlobalKey;
 
   @override
   _ContextMenuRouteStaticState createState() => _ContextMenuRouteStaticState();
@@ -921,13 +902,13 @@ class _ContextMenuRouteStaticState extends State<_ContextMenuRouteStatic>
   static const double _kDamping = 400.0;
   static const Duration _kMoveControllerDuration = Duration(milliseconds: 600);
 
-  Offset _dragOffset;
+  late Offset _dragOffset;
   double _lastScale = 1.0;
-  AnimationController _moveController;
-  AnimationController _sheetController;
-  Animation<Offset> _moveAnimation;
-  Animation<double> _sheetScaleAnimation;
-  Animation<double> _sheetOpacityAnimation;
+  late AnimationController _moveController;
+  late AnimationController _sheetController;
+  late Animation<Offset> _moveAnimation;
+  late Animation<double> _sheetScaleAnimation;
+  late Animation<double> _sheetOpacityAnimation;
 
   // The scale of the child changes as a function of the distance it is dragged.
   static double _getScale(
@@ -977,7 +958,7 @@ class _ContextMenuRouteStaticState extends State<_ContextMenuRouteStatic>
 
     // Dismiss if the drag is enough to scale down all the way.
     if (_lastScale == _kMinScale) {
-      widget.onDismiss(context, _lastScale, _sheetOpacityAnimation.value);
+      widget.onDismiss!(context, _lastScale, _sheetOpacityAnimation.value);
       return;
     }
 
@@ -1010,7 +991,7 @@ class _ContextMenuRouteStaticState extends State<_ContextMenuRouteStatic>
     if (_moveAnimation.value.dy == 0.0) {
       return;
     }
-    widget.onDismiss(context, _lastScale, _sheetOpacityAnimation.value);
+    widget.onDismiss!(context, _lastScale, _sheetOpacityAnimation.value);
   }
 
   Alignment _getChildAlignment(
@@ -1024,7 +1005,7 @@ class _ContextMenuRouteStaticState extends State<_ContextMenuRouteStatic>
         return orientation == Orientation.portrait
             ? Alignment.bottomCenter
             : Alignment.topLeft;
-      default:
+      case _ContextMenuLocation.left:
         return orientation == Orientation.portrait
             ? Alignment.bottomCenter
             : Alignment.topRight;
@@ -1042,7 +1023,7 @@ class _ContextMenuRouteStaticState extends State<_ContextMenuRouteStatic>
       _moveAnimation = Tween<Offset>(
         begin: Offset.zero,
         end: Offset(
-          endX.clamp(-_kPadding, _kPadding) as double,
+          endX.clamp(-_kPadding, _kPadding),
           endY,
         ),
       ).animate(
@@ -1079,10 +1060,7 @@ class _ContextMenuRouteStaticState extends State<_ContextMenuRouteStatic>
         child: AnimatedBuilder(
           animation: _moveController,
           builder: _buildChildAnimation,
-          child: AbsorbPointer(
-            absorbing: widget.absorbing,
-            child: widget.child,
-          ),
+          child: widget.child,
         ),
       ),
     );
@@ -1096,7 +1074,7 @@ class _ContextMenuRouteStaticState extends State<_ContextMenuRouteStatic>
         builder: _buildSheetAnimation,
         child: _ContextMenuSheet(
           key: widget.sheetGlobalKey,
-          actions: widget.actions,
+          actions: widget.actions!,
           contextMenuLocation: widget.contextMenuLocation,
           orientation: widget.orientation,
         ),
@@ -1110,13 +1088,13 @@ class _ContextMenuRouteStaticState extends State<_ContextMenuRouteStatic>
         return orientation == Orientation.portrait
             ? <Widget>[child, spacer, sheet]
             : <Widget>[sheet, spacer, child];
-      default:
+      case _ContextMenuLocation.left:
         return <Widget>[child, spacer, sheet];
     }
   }
 
   // Build the animation for the _ContextMenuSheet.
-  Widget _buildSheetAnimation(BuildContext context, Widget child) {
+  Widget _buildSheetAnimation(BuildContext context, Widget? child) {
     return Transform.scale(
       alignment:
           _ContextMenuRoute.getSheetAlignment(widget.contextMenuLocation),
@@ -1129,7 +1107,7 @@ class _ContextMenuRouteStaticState extends State<_ContextMenuRouteStatic>
   }
 
   // Build the animation for the child.
-  Widget _buildChildAnimation(BuildContext context, Widget child) {
+  Widget _buildChildAnimation(BuildContext context, Widget? child) {
     _lastScale = _getScale(
       widget.orientation,
       MediaQuery.of(context).size.height,
@@ -1143,7 +1121,7 @@ class _ContextMenuRouteStaticState extends State<_ContextMenuRouteStatic>
   }
 
   // Build the animation for the overall draggable dismissible content.
-  Widget _buildAnimation(BuildContext context, Widget child) {
+  Widget _buildAnimation(BuildContext context, Widget? child) {
     return Transform.translate(
       offset: _moveAnimation.value,
       child: child,
@@ -1227,10 +1205,10 @@ class _ContextMenuRouteStaticState extends State<_ContextMenuRouteStatic>
 // list of actions that are typically CupertinoContextMenuActions.
 class _ContextMenuSheet extends StatelessWidget {
   _ContextMenuSheet({
-    Key key,
-    @required this.actions,
-    @required _ContextMenuLocation contextMenuLocation,
-    @required Orientation orientation,
+    Key? key,
+    required this.actions,
+    required _ContextMenuLocation contextMenuLocation,
+    required Orientation orientation,
   })  : assert(actions != null && actions.isNotEmpty),
         assert(contextMenuLocation != null),
         assert(orientation != null),
@@ -1284,7 +1262,7 @@ class _ContextMenuSheet extends StatelessWidget {
           ),
           menu,
         ];
-      default:
+      case _ContextMenuLocation.left:
         return <Widget>[
           menu,
           const Spacer(
@@ -1310,11 +1288,11 @@ class _ContextMenuSheet extends StatelessWidget {
 // offColor at the times given by intervalOn and intervalOff.
 class _OnOffAnimation<T> extends CompoundAnimation<T> {
   _OnOffAnimation({
-    AnimationController controller,
-    @required T onValue,
-    @required T offValue,
-    @required double intervalOn,
-    @required double intervalOff,
+    required AnimationController controller,
+    required T onValue,
+    required T offValue,
+    required double intervalOn,
+    required double intervalOff,
   })  : _offValue = offValue,
         assert(intervalOn >= 0.0 && intervalOn <= 1.0),
         assert(intervalOff >= 0.0 && intervalOff <= 1.0),
